Added formatting to Views; stuck on why Edit View isn't populating

+++ Models/Pet.cs
@@ -30,7 +30,7 @@ public class Pet {
 
     [Required, Range(0, 30)]
     [Display(Name = "Pet Age")]
-    public int PetAge { get; set; }
+    public float PetAge { get; set; }
 
     [Required]
     [Display(Name = "Pet Gender")]
